adjusting the float type to have a longer decimal value for lat lng

+++ models/geo.js
@@ -8,11 +8,11 @@ module.exports = function (sequelize, DataTypes) {
       allowNull: false,
     },
     lat: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.FLOAT(11,7),
       allowNull: false,
     },
     lng: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.FLOAT(11,7),
       allowNull: false,
     },
   });
